scoping by a database field
In this commit, we will use again the foreign key scoping instead of
schema scoping. Both have their pros and cons.
With PostgreSQL scoping, the scoping happens implicitly, by the
Apartment gem in conjunction with PostgreSQL’s schema search path, and
so there is no need for putting scoping calls everywhere in the
application. However, the downside to this is that it will only work on
small applications with a small number of tables. Applications with a
large amount of total tables may have problems running backups.
With foreign key scoping, it needs to make explicit calls to scope your
application’s resources wherever that is happening. Missing even a
single place can lead to the application not behaving as intended.